+++ minecraft_server/net/minecraft/entity/ai/RandomPositionGenerator.java
@@ -65,8 +65,8 @@ public class RandomPositionGenerator
 
         if (par0EntityCreature.hasHome())
         {
-            double var11 = (double)(par0EntityCreature.getHomePosition().getDistanceSquared(MathHelper.floor_double(par0EntityCreature.posX), MathHelper.floor_double(par0EntityCreature.posY), MathHelper.floor_double(par0EntityCreature.posZ)) + 4.0F);
-            double var13 = (double)(par0EntityCreature.func_110174_bM() + (float)par1);
+            double var11 = par0EntityCreature.getHomePosition().getDistanceSquared(MathHelper.floor_double(par0EntityCreature.posX), MathHelper.floor_double(par0EntityCreature.posY), MathHelper.floor_double(par0EntityCreature.posZ)) + 4.0F;
+            double var13 = par0EntityCreature.func_110174_bM() + par1;
             var10 = var11 < var13 * var13;
         }
         else
@@ -80,7 +80,7 @@ public class RandomPositionGenerator
             int var17 = var4.nextInt(2 * par2) - par2;
             int var14 = var4.nextInt(2 * par1) - par1;
 
-            if (par3Vec3 == null || (double)var12 * par3Vec3.xCoord + (double)var14 * par3Vec3.zCoord >= 0.0D)
+            if (par3Vec3 == null || var12 * par3Vec3.xCoord + var14 * par3Vec3.zCoord >= 0.0D)
             {
                 var12 += MathHelper.floor_double(par0EntityCreature.posX);
                 var17 += MathHelper.floor_double(par0EntityCreature.posY);
@@ -104,7 +104,7 @@ public class RandomPositionGenerator
 
         if (var5)
         {
-            return par0EntityCreature.worldObj.getWorldVec3Pool().getVecFromPool((double)var6, (double)var7, (double)var8);
+            return par0EntityCreature.worldObj.getWorldVec3Pool().getVecFromPool(var6, var7, var8);
         }
         else
         {
